Version 0.18.0: uint -> int64.

+++ internal/application/chat/dto.go
@@ -1,14 +1,14 @@
 package chat
 
 type MessageCreateDTO struct {
-	AuthorID    uint   `json:"author_id"`
+	AuthorID    int64  `json:"author_id"`
 	Body        string `json:"body"`
 	MessageType string `json:"type"`
 }
 
 type MessageDTO struct {
-	ID          uint   `json:"id"`
-	AuthorID    uint   `json:"author_id"`
+	ID          int64  `json:"id"`
+	AuthorID    int64  `json:"author_id"`
 	Username    string `json:"username"`
 	Body        string `json:"body"`
 	SentAt      string `json:"sent_at"`
