+++ internal/domain/questionRepo.go
@@ -4,7 +4,7 @@ type QuestionRepository interface {
 	All() ([]*Question, error)
 	DeleteAll() error
 	Random() (*Question, error)
-	GetByID(id uint) (*Question, error)
-	RandomID() (uint, error)
+	GetByID(id int64) (*Question, error)
+	RandomID() (int64, error)
 	FindByTopicID(topicID uint) ([]*Question, error)
 }
